Reorganize the `suballocator` module
